src/requirements.txt
# Flask app
flask==2.3.3

# Firestore integration
google-cloud-firestore==2.21.0
google-auth==2.40.3
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Testes
pytest==8.4.1

flask>=2.0.0
gunicorn>=20.0.0
google-cloud-firestore>=2.0.0

openai>=1.0.0
python-dotenv
